fastapi/requirements.txt
# app
fastapi
uvicorn
motor
python-dotenv
pyaml
# test
pytest
httpx
requests